Fix test user service problem

+++ src/main/java/hr/prism/board/service/TestUserService.java
@@ -1,7 +1,7 @@
 package hr.prism.board.service;
 
-import com.google.common.collect.ImmutableList;
 import hr.prism.board.repository.UserRepository;
+import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -10,11 +10,13 @@ import javax.persistence.EntityManager;
 import javax.persistence.Query;
 import java.util.List;
 
+import static org.slf4j.LoggerFactory.getLogger;
+
 @Service
 @Transactional
 public class TestUserService {
 
-    private static final ImmutableList<String> FRAMEWORK_TABLES = ImmutableList.of("flyway_schema_history", "workflow");
+    private static final Logger LOGGER = getLogger(TestUserService.class);
 
     private final UserRepository userRepository;
 
@@ -30,26 +32,33 @@ public class TestUserService {
     public void deleteTestUsers() {
         List<Long> userIds = userRepository.findByTestUser(true);
         if (!userIds.isEmpty()) {
-            Query removeForeignKeyChecks =
-                entityManager.createNativeQuery("SET SESSION FOREIGN_KEY_CHECKS = 0");
-            removeForeignKeyChecks.executeUpdate();
+            entityManager.createNativeQuery("SET SESSION FOREIGN_KEY_CHECKS = 0")
+                .executeUpdate();
 
-            List<String> tablesNames = entityManager.createNativeQuery("SHOW TABLES").getResultList();
+            List<String> tablesNames =
+                entityManager.createNativeQuery("SHOW TABLES")
+                    .getResultList();
 
             tablesNames
-                .stream()
-                .filter(tableName -> !FRAMEWORK_TABLES.contains(tableName))
                 .forEach(tableName -> {
-                    Query deleteUserData = entityManager.createNativeQuery(
-                        "DELETE FROM " + tableName + " WHERE creator_id IN (:userIds)");
+                    List<Object[]> columns =
+                        entityManager.createNativeQuery("SHOW COLUMNS FROM " + tableName)
+                            .getResultList();
 
-                    deleteUserData.setParameter("userIds", userIds);
-                    deleteUserData.executeUpdate();
+                    if (columns.stream()
+                        .anyMatch(column -> "creator_id".equals(column[0]))) {
+                        LOGGER.info("Deleting test users from table: " + tableName);
+
+                        Query deleteUserData = entityManager.createNativeQuery(
+                            "DELETE FROM " + tableName + " WHERE creator_id IN (:userIds)");
+
+                        deleteUserData.setParameter("userIds", userIds);
+                        deleteUserData.executeUpdate();
+                    }
                 });
 
-            Query restoreForeignKeyChecks = entityManager.createNativeQuery(
-                "SET SESSION FOREIGN_KEY_CHECKS = 1");
-            restoreForeignKeyChecks.executeUpdate();
+            entityManager.createNativeQuery("SET SESSION FOREIGN_KEY_CHECKS = 1")
+                .executeUpdate();
         }
     }
